Refactor project structure: Merge services into monolith
- Remove legacy `services/mini-redis-core` and `services/semantic-cache`
- Move core logic to `src/lib` and `src/routes`
- Update `package.json` to be single-package (remove workspaces)
- Update `docker-compose.yml` to use single `erion-ember` service
- Integrate `SemanticCache` into `chat` route
- Update tests imports
