--- core/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@erion/core",
-  "version": "1.0.0",
-  "description": "Semantic Cache Core with HTTP API",
-  "main": "src/server.js",
-  "scripts": {
-    "dev": "bun run --watch src/server.js",
-    "start": "bun run src/server.js",
-    "test": "bun test",
-    "build": "echo 'No build step needed for Bun'"
-  },
-  "dependencies": {
-    "fastify": "^4.24.0",
-    "@fastify/cors": "^8.4.0",
-    "ioredis": "^5.3.0",
-    "zod": "^3.22.0"
-  },
-  "devDependencies": {
-    "bun-types": "latest"
-  }
-}
