--- core/src/cache/index.js
@@ -1,6 +1,0 @@
-export { SemanticCache } from './semantic-cache.js';
-export { HNSWIndex } from './hnsw-index.js';
-export { Quantizer } from './quantizer.js';
-export { Normalizer } from './normalizer.js';
-export { Compressor } from './compressor.js';
-export { MetadataStore } from './metadata-store.js';
