--- services/mini-redis-core/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "mini-redis-core",
-  "version": "1.0.0",
-  "description": "High-performance Redis-compatible core server - lightweight and optimized for maximum throughput",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "NODE_ENV=development node index.js",
-    "prod": "NODE_ENV=production node index.js"
-  },
-  "keywords": [
-    "redis",
-    "in-memory",
-    "database",
-    "pubsub",
-    "performance",
-    "tcp-server",
-    "high-throughput",
-    "microservice"
-  ],
-  "author": "EricNguyen1206",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/EricNguyen1206/mini-redis",
-    "directory": "services/mini-redis-core"
-  },
-  "engines": {
-    "node": ">=14.0.0"
-  },
-  "dependencies": {},
-  "devDependencies": {},
-  "bin": {
-    "mini-redis-core": "./index.js"
-  }
-}
