--- services/semantic-cache/.gitignore
@@ -1,7 +1,0 @@
-node_modules/
-coverage/
-*.log
-.DS_Store
-data/
-*.bin
-*.snapshot
