--- services/semantic-cache/CHANGELOG.md
@@ -1,171 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [1.0.0] - 2026-01-31
-
-### Added
-
-#### Core Components
-
-- **Vector Quantizer** (`src/quantizer.js`)
-  - FP32 to INT8 quantization for 75% memory reduction
-  - Maps float range [-1, 1] to integer range [0, 255]
-  - <2% accuracy loss in cosine similarity
-  - Round-trip quantization and dequantization support
-
-- **LZ4 Compressor** (`src/compressor.js`)
-  - Fast text compression for LLM responses
-  - 60-80% size reduction for typical content
-  - Compression level 1 for maximum speed
-  - ~500MB/s decompression throughput
-
-- **Prompt Normalizer** (`src/normalizer.js`)
-  - Text normalization: lowercase, trim, collapse whitespace
-  - SHA256 hashing for deduplication using Node.js crypto
-  - Consistent cache keys for similar prompts
-
-- **HNSW Index Wrapper** (`src/hnsw-index.js`)
-  - C++ vector search via hnswlib-node bindings
-  - Approximate Nearest Neighbor (ANN) with O(log n) complexity
-  - Cosine similarity metric optimized for embeddings
-  - Configurable: M=16, efConstruction=200, ef=100
-  - Synchronous save/load with `writeIndexSync`/`readIndexSync`
-
-- **Metadata Store** (`src/metadata-store.js`)
-  - LRU (Least Recently Used) eviction policy
-  - JavaScript Map-based storage for O(1) lookups
-  - Prompt hash indexing for exact match detection
-  - Access tracking: count, timestamps, compression stats
-  - Configurable maximum cache size
-
-- **Semantic Cache Core** (`src/semantic-cache.js`)
-  - Main API combining all components
-  - Exact match detection via hash lookup
-  - Semantic search via HNSW vector similarity
-  - Configurable similarity threshold (default: 0.85)
-  - Statistics tracking: hits, misses, memory usage
-  - Persistence: save/load cache to disk
-
-- **Server Entry Point** (`src/index.js`)
-  - CLI server with environment variable configuration
-  - Graceful shutdown handling (SIGINT/SIGTERM)
-  - Periodic statistics reporting (30s interval)
-  - Support for CACHE_DIM, CACHE_MAX_ELEMENTS, CACHE_THRESHOLD, CACHE_PORT
-
-#### Infrastructure & Tooling
-
-- **Docker Support**
-  - Dockerfile with build-essential for native C++ modules
-  - docker-compose.yml with development and test services
-  - Volume mounting for live code changes
-  - Node.js 20 (Bullseye) base image
-
-- **Testing Framework**
-  - Jest test runner with 23 unit tests
-  - Test coverage for all core components
-  - Docker-based test execution
-  - Test commands: `npm test`, `npm run test:watch`, `npm run test:coverage`
-
-- **Project Structure**
-  - Organized src/ and test/ directories
-  - CommonJS module system
-  - MIT license
-  - Comprehensive README.md
-
-### Technical Achievements
-
-#### Memory Efficiency
-- **75% reduction** in vector storage (FP32 → INT8)
-- **60-80% reduction** in response storage (LZ4 compression)
-- **Overall 60-70% memory savings** compared to raw Redis storage
-- Example: 100K entries with 1536-dim vectors
-  - Raw: ~1 GB
-  - Compressed: ~350 MB
-
-#### Performance Characteristics
-- Query latency: <10ms (P95) for cache hits
-- Throughput: 5,000+ queries per second
-- Cache hit rate: >80% for semantically similar prompts
-- HNSW search: O(log n) complexity with 95%+ recall
-
-### Dependencies
-
-#### Production
-- `hnswlib-node`: ^2.0.0 - C++ HNSW implementation
-- `lz4`: ^0.6.0 - Fast compression algorithm
-- `xxhash-addon`: ^2.0.0 - Fast hashing (note: replaced with crypto in implementation)
-
-#### Development
-- `jest`: ^29.7.0 - Testing framework
-
-### Known Issues
-
-- `xxhash-addon` causes segmentation fault on some systems, replaced with Node.js built-in `crypto` module
-- HNSW index requires synchronous file operations (async API not stable)
-- Docker required for development due to C++ compilation dependencies
-
-## [Unreleased]
-
-### Planned Features
-
-- TTL (Time To Live) support for cache entries
-- Multi-modal embeddings support (images, audio)
-- Distributed cache backend (Redis/Valkey integration)
-- GPU acceleration for vector operations
-- Prometheus metrics export
-- REST API endpoints for HTTP access
-- WebSocket support for real-time updates
-- Authentication and rate limiting
-- Multi-tenant namespace isolation
-
-### Planned Improvements
-
-- Product Quantization (PQ) for 10-20x compression
-- Graph-based search with knowledge graph integration
-- Adaptive similarity thresholds
-- Predictive cache warming
-- Compression algorithm selection (LZ4/Zstd/Snappy)
-
----
-
-## Migration Guide
-
-### From v0.x to v1.0.0
-
-This is the initial release. No migration needed.
-
-### Environment Setup
-
-```bash
-# Install dependencies
-cd services/semantic-cache
-npm install
-
-# Run tests
-docker-compose -f docker-compose.yml run semantic-cache npm test
-
-# Start development server
-docker-compose -f docker-compose.yml up semantic-cache-dev
-```
-
-### Configuration
-
-Environment variables:
-- `CACHE_DIM`: Embedding dimension (default: 1536)
-- `CACHE_MAX_ELEMENTS`: Maximum cache entries (default: 100000)
-- `CACHE_THRESHOLD`: Similarity threshold 0-1 (default: 0.85)
-- `CACHE_PORT`: Server port (default: 3000)
-
----
-
-## Contributors
-
-- Initial implementation by AI Assistant (2026-01-31)
-
-## License
-
-MIT License - see LICENSE file for details
