--- services/semantic-cache/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "semantic-cache",
-  "version": "1.0.0",
-  "description": "High-performance semantic cache for LLM queries with HNSW vector search",
-  "main": "src/index.js",
-  "scripts": {
-    "start": "node src/index.js",
-    "dev": "NODE_ENV=development node src/index.js",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
-  },
-  "keywords": [
-    "semantic-cache",
-    "llm",
-    "vector-search",
-    "hnsw",
-    "embedding"
-  ],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "hnswlib-node": "^2.0.0",
-    "lz4": "^0.6.0",
-    "xxhash-addon": "^2.0.0"
-  },
-  "devDependencies": {
-    "jest": "^29.7.0"
-  },
-  "engines": {
-    "node": ">=18.0.0"
-  }
-}
