+++ src/lib/compressor.js
