+++ src/lib/hnsw-index.js
