+++ src/lib/metadata-store.js
