+++ src/lib/normalizer.js
