+++ src/lib/quantizer.js
