+++ src/lib/semantic-cache.js
